[*] new pom for generic content card functions, and moved from forumPage
